Move fixtures to shared conftest for re-use
These same fixtures will be useful within the update_images tests
especially when we split the hourly function into hourly and
cloud access.
